backend/requirements.txt
flask
flask-cors
playwright
sentence-transformers
requests
python-dotenv
aiohttp
gunicorn